Add files via upload
My shitty script for PCA + Symbolic regression. This contains multiple blocks, which are commented. Block 0: performs PCA on log attenuation curves, set up to loop on 1 to 5 PC, checking what is the ideal number of components. Block 1: Again, PCA is now on 3 components (which seems the best option based on block 0) and uses a gplearn symbolic regressor to try to predict these PCs based on galaxy properties. These two blocks are commented. The currently uncommented block is PCA + PySr to try to model the PCs based on galaxy properties. Here we use a subsample of the initial attenuation curves as PySr is very slow.